Node check restructuring
