+++ backend/src/api/__init__.py
@@ -2,6 +2,7 @@ from .api import *
 from .group import *
 from .input import *
 from .node_context import *
+from .node_data import *
 from .output import *
 from .settings import *
 from .types import *
