+++ backend/src/api/node_check.py
@@ -8,9 +8,8 @@ from collections import OrderedDict
 from enum import Enum
 from typing import Any, Callable, NewType, Tuple, Union, cast, get_args
 
-from .input import BaseInput
 from .node_context import NodeContext
-from .output import BaseOutput
+from .node_data import NodeData
 
 _Ty = NewType("_Ty", object)
 
@@ -148,11 +147,8 @@ def get_type_annotations(fn: Callable) -> dict[str, _Ty]:
     return type_annotations
 
 
-def validate_return_type(return_type: _Ty, outputs: list[BaseOutput]):
-    if str(return_type).startswith("api.Iterator["):
-        return_type = get_args(return_type)[0]
-    elif str(return_type).startswith("api.Collector["):
-        return_type = get_args(return_type)[1]
+def validate_return_type(return_type: _Ty, node: NodeData):
+    outputs = node.outputs
 
     if len(outputs) == 0:
         if return_type is not None and return_type is not type(None):  # type: ignore
@@ -190,19 +186,20 @@ def validate_return_type(return_type: _Ty, outputs: list[BaseOutput]):
 
 def check_schema_types(
     wrapped_func: Callable,
-    inputs: list[BaseInput],
-    outputs: list[BaseOutput],
-    node_context: bool,
+    node: NodeData,
 ):
     """
     Runtime validation for the number of inputs/outputs compared to the type args
     """
 
+    if node.kind != "regularNode":
+        return
+
     ann = OrderedDict(get_type_annotations(wrapped_func))
 
     # check return type
     if "return" in ann:
-        validate_return_type(ann.pop("return"), outputs)
+        validate_return_type(ann.pop("return"), node)
 
     # check arguments
     arg_spec = inspect.getfullargspec(wrapped_func)
@@ -210,7 +207,7 @@ def check_schema_types(
         if arg not in ann:
             raise CheckFailedError(f"Missing type annotation for '{arg}'")
 
-    if node_context:
+    if node.node_context:
         first = arg_spec.args[0]
         if first != "context":
             raise CheckFailedError(
@@ -223,6 +220,7 @@ def check_schema_types(
             )
 
     # check inputs
+    inputs = node.inputs
 
     if arg_spec.varargs is not None:
         if arg_spec.varargs not in ann:
